style: 리팩토링 이전 파일 삭제 후 리팩토링 파일로 대체
